+++ src/mkvtoolnix-gui/main_window/preferences_dialog.cpp
@@ -88,6 +88,7 @@ PreferencesDialog::PreferencesDialog(QWidget *parent,
   if (!m_cfg.m_mediaInfoExe.isEmpty())
     ui->leMMediaInfoExe->setText(QDir::toNativeSeparators(m_cfg.m_mediaInfoExe));
   ui->cbMSortFilesTracksByTypeWhenAdding->setChecked(m_cfg.m_mergeSortFilesTracksByTypeWhenAdding);
+  ui->cbMReconstructSequencesWhenAdding->setChecked(m_cfg.m_mergeReconstructSequencesWhenAdding);
   ui->cbMAutoSetFileTitle->setChecked(m_cfg.m_autoSetFileTitle);
   ui->cbMAutoClearFileTitle->setChecked(m_cfg.m_autoClearFileTitle);
   ui->cbMSetAudioDelayFromFileName->setChecked(m_cfg.m_setAudioDelayFromFileName);
@@ -368,6 +369,15 @@ PreferencesDialog::setupToolTips() {
                    .arg(QY("Similarly an audio track will be inserted before the first track that's neither a video nor an audio track."))
                    .arg(QY("If disabled, files and tracks will be inserted after all existing files and tracks."))
                    .arg(QY("This only determines the initial order which can still be changed manually later.")));
+
+  Util::setToolTip(ui->cbMReconstructSequencesWhenAdding,
+                   Q("<p>%1 %2 %3</p><p>%4</p><p>%5</p><p>%6</p>")
+                   .arg(QY("If enabled, the GUI will analyze the file names when you add multiple files at once."))
+                   .arg(QY("It tries to detect sequences of names that likely belong to the same movie by splitting the name into three parts: a prefix, a running number and a suffix that doesn't contain digits."))
+                   .arg(QY("Names are considered to be in sequence when the previous file name's prefix & suffix match the current file name's prefix & suffix and the running number is incremented by one."))
+                   .arg(QY("An example: 'movie.001.mp4', 'movie.002.mp4' and 'movie.003.mp4'"))
+                   .arg(QY("If such a sequence is detected, only the first file is added while the second and following files in the sequence are appended to the first one."))
+                   .arg(QY("If disabled, all files selected for adding will always be added regardless of their names.")));
 
   Util::setToolTip(ui->cbMAlwaysShowOutputFileControls,
                    Q("%1 %2")
@@ -937,6 +947,7 @@ PreferencesDialog::save() {
   m_cfg.m_mergeDragAndDropFilesPolicy                         = static_cast<Util::Settings::MergeAddingAppendingFilesPolicy>(ui->cbMDragAndDropFilesPolicy->currentData().toInt());
   m_cfg.m_mergeAlwaysCreateNewSettingsForVideoFiles           = ui->cbMAlwaysCreateSettingsForVideoFiles->isChecked();
   m_cfg.m_mergeSortFilesTracksByTypeWhenAdding                = ui->cbMSortFilesTracksByTypeWhenAdding->isChecked();
+  m_cfg.m_mergeReconstructSequencesWhenAdding                 = ui->cbMReconstructSequencesWhenAdding->isChecked();
   m_cfg.m_mergeAlwaysShowOutputFileControls                   = ui->cbMAlwaysShowOutputFileControls->isChecked();
   m_cfg.m_mergeWarnMissingAudioTrack                          = static_cast<Util::Settings::MergeMissingAudioTrackPolicy>(ui->cbMWarnMissingAudioTrack->currentData().toInt());
   m_cfg.m_mergeTrackPropertiesLayout                          = ui->rbMTrackPropertiesLayoutHorizontalScrollArea->isChecked() ? Util::Settings::TrackPropertiesLayout::HorizontalScrollArea