+++ src/mkvtoolnix-gui/merge/source_file_model.cpp
@@ -1,11 +1,12 @@
 #include "common/common_pch.h"
 
+#include <QByteArray>
+#include <QDataStream>
 #include <QDir>
 #include <QFileInfo>
 #include <QItemSelectionModel>
 #include <QMimeData>
-#include <QByteArray>
-#include <QDataStream>
+#include <QRegularExpression>
 
 #include "common/logger.h"
 #include "common/sorting.h"
@@ -21,6 +22,28 @@
 namespace mtx::gui::Merge {
 
 namespace {
+
+struct SequencedFileNameData {
+  QString prefix, suffix;
+  unsigned int number{};
+
+  bool follows(SequencedFileNameData const &previous) const {
+    return (prefix == previous.prefix)
+        && (suffix == previous.suffix)
+        && (number == (previous.number + 1));
+  }
+};
+
+std::optional<SequencedFileNameData>
+analyzeFileNameForSequenceData(QString const &fileName) {
+  QRegularExpression re{Q(R"(([^/\\]*)(\d+)([^\d]+)$)")};
+  auto match = re.match(fileName);
+
+  if (match.hasMatch())
+    return SequencedFileNameData{ match.captured(1), match.captured(3), match.captured(2).toUInt() };
+
+  return {};
+}
 
 int
 insertPriorityForSourceFile(SourceFile const &file) {
@@ -261,7 +284,7 @@ SourceFileModel::addOrAppendFilesAndTracks(QVector<SourceFilePtr> const &files,
     addFilesAndTracks(files);
 }
 
-bool
+QModelIndex
 SourceFileModel::addFileSortedByType(SourceFilePtr const &file) {
   auto newFilePrio = insertPriorityForSourceFile(*file);
 
@@ -277,28 +300,61 @@ SourceFileModel::addFileSortedByType(SourceFilePtr const &file) {
     invisibleRootItem()->insertRow(idx, row);
     m_sourceFiles->insert(idx, file);
 
-    return true;
+    return index(idx, 0);
   }
 
-  return false;
+  return {};
 }
 
-void
+QModelIndex
 SourceFileModel::addFileAtAppropriatePlace(SourceFilePtr const &file,
                                            bool sortByType) {
-  if (sortByType && addFileSortedByType(file))
-    return;
+  QModelIndex insertPosIdx;
+
+  if (sortByType) {
+    insertPosIdx = addFileSortedByType(file);
+
+    if (insertPosIdx.isValid())
+      return insertPosIdx;
+  }
 
   createAndAppendRow(invisibleRootItem(), file);
   *m_sourceFiles << file;
+
+  return index(rowCount() - 1, 0);
 }
 
 void
 SourceFileModel::addFilesAndTracks(QVector<SourceFilePtr> const &files) {
-  auto sortByType = Util::Settings::get().m_mergeSortFilesTracksByTypeWhenAdding;
+  std::optional<SequencedFileNameData> previouslyAddedSequenceData;
+  QModelIndex previouslyAddedPosition;
 
-  for (auto const &file : files) {
-    addFileAtAppropriatePlace(file, sortByType);
+  auto &cfg                 = Util::Settings::get();
+  auto sortByType           = cfg.m_mergeSortFilesTracksByTypeWhenAdding;
+  auto reconstructSequences = cfg.m_mergeReconstructSequencesWhenAdding;
+  auto filesToProcess       = files;
+
+  if (reconstructSequences)
+    std::sort(filesToProcess.begin(), filesToProcess.end(), [](auto const &a, auto const &b) {
+        return a->m_fileName < b->m_fileName;
+      });
+
+  for (auto const &file : filesToProcess) {
+    auto sequenceData = analyzeFileNameForSequenceData(file->m_fileName);
+
+    if (   reconstructSequences
+        && previouslyAddedSequenceData
+        && previouslyAddedPosition.isValid()
+        && sequenceData
+        && sequenceData->follows(*previouslyAddedSequenceData)) {
+      appendFilesAndTracks({ file }, previouslyAddedPosition);
+      previouslyAddedSequenceData = sequenceData;
+
+      continue;
+    }
+
+    previouslyAddedPosition     = addFileAtAppropriatePlace(file, sortByType);
+    previouslyAddedSequenceData = sequenceData;
 
     for (auto const &track : file->m_tracks)
       m_tracksModel->addTrackAtAppropriatePlace(track, sortByType);