+++ NEWS.md
@@ -15,6 +15,12 @@
   first followed by audio, subtitles and other types. Files & tracks can still
   be reordered manually later. This option is enabled by default. Implements
   #2366.
+* MKVToolNix GUI: multiplexer: added an option for recognizing file name
+  sequences such as `movie.001.mp4`, `movie.002.mp4`, `movie.003.mp4` when
+  adding multiple files at once. If a sequence is detected, the only first
+  file will be added while the second and following file names will be
+  appended to the first one. This option is enabled by default. Implements
+  #2866.
 
 ## Bug fixes